refactor(anyharness): extract npm platform resolution to its own module
installer/npm.rs reached 636 lines against the 600-line PROD-SIZE-1 cap.
lints/anyharness/ratchets.toml is a ratchet, not an exception ledger, and its
own header says entries may only shrink, so adding a new entry for growth this
PR introduced is the wrong move. Shrink the file instead.

Pure move: npm_program_name{,_windows,_unix}, platform_npm_bin_relpath{,_windows,_unix}
and their four tests go to installer/npm_platform.rs. npm.rs re-exports
platform_npm_bin_relpath as pub(super) so agent_process.rs's existing
'use super::npm::{...}' import keeps resolving unchanged, which also keeps
PR #2153's verified merge resolution valid.

npm.rs 636 -> 535. check_max_lines.py: 'Max-lines check passed'.

Negative control (mutate npm.cmd -> npm and drop the .cmd suffix):
  2 failed; 2 passed
restored:
  test result: ok. 4 passed; 0 failed
